120% last commit that changing auth pages.

+++ src/middleware.ts
@@ -7,13 +7,14 @@ import {
 } from "./constants/api/jwt";
 import {
   API_ROUTE,
+  CHATS_ROUTE,
   DASHBOARD_ROUTE,
   LOGIN_ROUTE,
   LOGOUT_ROUTE,
   SIGNUP_ROUTE,
 } from "./constants/routes";
 import JWTTokenService from "./services/server/JWTTokenService";
-import { SESSION_ROUTE } from "./constants/api/routes";
+import { SESSION_API_ROUTE } from "./constants/api/routes";
 
 const jwtService = new JWTTokenService();
 
@@ -30,15 +31,17 @@ export async function middleware(request: NextRequest) {
     return nextResponse;
   }
 
+  if (checkRoute === DASHBOARD_ROUTE) {
+    return NextResponse.redirect(new URL(CHATS_ROUTE, request.url));
+  }
+
   if (checkRoute.startsWith(LOGOUT_ROUTE)) {
-    const nextResponse = NextResponse.redirect(
-      new URL(LOGIN_ROUTE, request.url),
-    );
+    const nextResponse = NextResponse.next();
     nextResponse.cookies.delete(REFRESH_TOKEN);
     nextResponse.cookies.delete(ACCESS_TOKEN);
     nextResponse.cookies.delete(ATTEMPTED_REFRESH);
 
-    await fetch(SESSION_ROUTE, {
+    await fetch(SESSION_API_ROUTE, {
       method: "DELETE",
       headers: {
         "Content-type": "application/json",
@@ -71,7 +74,7 @@ export async function middleware(request: NextRequest) {
     );
 
     if (verifyAccessTokenError) {
-      const response = await fetch(SESSION_ROUTE, {
+      const response = await fetch(SESSION_API_ROUTE, {
         method: "PUT",
         headers: {
           "Content-type": "application/json",
